+++ cnccode/controller_classes/CTCustomer.inc.php
@@ -908,36 +908,6 @@ class CTCustomer extends CTCNC
         return $this->newCustomerToDate;
     }
 
-    function setReviewMeetingFrequencyMonths($reviewMeetingFrequencyMonths)
-    {
-        $this->reviewMeetingFrequencyMonths = $reviewMeetingFrequencyMonths;
-    }
-
-    function getReviewMeetingFrequencyMonths()
-    {
-        return $this->reviewMeetingFrequencyMonths;
-    }
-
-    function setLastReviewMeetingDate($lastReviewMeetingDate)
-    {
-        $this->lastReviewMeetingDate = $lastReviewMeetingDate;
-    }
-
-    function getLastReviewMeetingDate()
-    {
-        return $this->lastReviewMeetingDate;
-    }
-
-    function setReviewMeetingEmailSentFlag($value)
-    {
-        $this->reviewMeetingEmailSentFlag = $value;
-    }
-
-    function getReviewMeetingEmailSentFlag()
-    {
-        return $this->reviewMeetingEmailSentFlag;
-    }
-
     function setDroppedCustomerFromDate($droppedCustomerFromDate)
     {
         $this->droppedCustomerFromDate = $droppedCustomerFromDate;