+++ cnccode/data_classes/DBECustomer.inc.php
@@ -320,12 +320,6 @@ class DBECustomer extends DBCNCEntity
             'cus_review_meeting_frequency_months'
         );
         $this->addColumn(
-            self::lastReviewMeetingDate,
-            DA_DATE,
-            DA_ALLOW_NULL,
-            'cus_last_review_meeting_date'
-        );
-        $this->addColumn(
             self::reviewMeetingEmailSentFlag,
             DA_YN,
             DA_ALLOW_NULL,
@@ -744,6 +738,23 @@ class DBECustomer extends DBCNCEntity
         $this->setQueryString($queryString);
         $ret = (self::getRows());
         return $ret;
+    }
+
+    function getReviewMeetingCustomers()
+    {
+        $this->setMethodName('getReviewMeetingCustomers');
+        $queryString =
+            "SELECT " . $this->getDBColumnNamesAsString() .
+            " FROM " . $this->getTableName() .
+            " WHERE " . $this->getDBColumnName(self::lastReviewMeetingDate) . " <> '0000-00-00' and
+             " . $this->getDBColumnName(self::lastReviewMeetingDate) . ' is not null and ' .
+            $this->getDBColumnName(self::referredFlag) . ' = "N" ';
+
+        $this->setQueryString($queryString);
+        $ret = (self::getRows());
+        return $ret;
+
+
     }
 }