+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project changes will be shown here.
 - Issue 413 - Default home page graphs to your team
 - Issue 421 - Increase Questionnaire field length
 - Issue 422 - Decimal places for fix times in System Header
+- Issue 423 - Upcoming Customer Review Meeting
 ### Fixed
 - Issue 420 - Email missing header information
 - Issue 378 - Creating item defaults to first site